DirectorSyncV3/requirements.txt
PyYAML>=6,<7